requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
diff-match-patch==20241021
Django==5.0.6
django-cotton==1.3.0
django-environ==0.11.2
django-import-export==4.2.1
django-unfold==0.34.0
ldjango==8.5
prompt_toolkit==3.0.48
sqlparse==0.5.2
tablib==3.7.0
tqdm==4.67.0
wcwidth==0.2.13
whitenoise==6.8.2
